Verson 3 Upgrade
Update collivery.php - Updated version number of application as well as the WooCommerce version tested up to value.

Update _parcel_row.php - It was showing a warning due to it being deprecated.

Fixed issue where town saves as ID instead of name.

Updated MDS Confirmed page, so that Waybill and POD are visible. Risk Cover is also pulling through correctly.

Updated Collivery Error messages, so that they're logged correctly.

Error when no document was found. It would still carry on as if a document was found.

Updated code to handle POD's being returned when trying to view them on the MDS Confirmed screen.

Fixing issues related to the MDS Confirmed Page and Backward Compability. Also fixed an issue where the status only loaded the 2nd time and not the first time.

Cleaned up commented code.

Updated an issue with the Collivery object not being initialized.

Fixed a search by suburb method that was incorrectly getting the suburbs.

Fixing the prices to handle VAT accordingly.

Updated MDS Confirmed label from MDS Rate to MDS Rate Ex Vat.  Updated the quoted /actual weight/price. Updated back end to get the quoted price and weight.

Fixed most of the MDS Confirmed screen. Everything is being displayed. Minus the View PDF, View POD and 2 sections that need further clarification.

Updated the base page MDS Confirmed page to display correctly. Also realized I was saving the data in  customer_id instead of customer_reference. Updated this too.

Fixing the display of the Add Collivery Page, fixing up the getQuote functionality as well as the Accept/Dispatch button. Which includes adding an address/contact combo.

Fixed issue with the city appearing as an id on the Purchase details screen.

Fixing the checkout screen not working. Fixed the suburbs not populating when the page is reloaded.
Fixed the cart displaying an ID instead of the city name when giving an estimate on the cart.

The checkout screen correctly displays the shipping methods.

Fixed issues with caching, updated method of sending data to json. Updated the shipping estimation values to send the data in the correct format.

Fixed dropdown lists on checkout screen not being populated correctly.

I've updated nearly all of the API calls in Collivery.php to use new API. Created a new generic method to consume the new API. I've gotten the initial settings page working as well as the Shipping Zones for MDS Collivery.

Updated API URL, as well as Authenticate method