CP-179
Updated Array initializations to match proper syntax.

Made some changes to some of the exceptions thrown to return a little more information so it's easy to see what the issue is.

Fixed an issue where towns weren't converted correctly.

Implemented functionality to correctly use the custom_id when saving an address.