CP-225
script.js - Updated to hide/display fields on the checkout screen depending on which country is selected.

ShippingPackageData.php - Updated to set the country correctly based on the country selected instead of hardcoding to ZA. Also made some changes in regards to try and get the 'ship_to_different_address' to work. Not working.

MdsColliveryService.php - Updated the getPrice function so that when the $fixedPrice is populated it returns it immediately instead of doing the API call.

MdsCheckoutFields.php - Added some new classes to hide new fields. Also updated priorties and removed some content that wasn't being used.

Collivery.php - Added things that need to be removed once it goes live. Namely the SSL_VerifyPeer option as well as function for easy debugging (logError).

mds_checkout_fields.php - Removed the default value being sent through, as the getTowns function already defaults to "ZAF" if nothing is sent.