CP-242
script.js - Changes to ensure, that non South African cities and suburbs pull through correctly when Flat Rates all pulled out. As well as a method to remove them if the user were to swap back to South Africa.

ShippingPackageData.php - Removing this, caused issues. So I have added it back in, but not updating the Country. This removes warnings we got on the checkout screen when the shipping options were retreived.

MdsColliveryService.php - The $towns variable is null, this ensure we just return the name and stops it from crashing.